Recent games should work as well as ranked.
This is untested at the moment, but the structure is there to make the
client calls see the same data whether or not it is coming from the the
recent games or ranked (matchhistory) end points.

It now returns the team and champion id that the subject used in the
game.